Precautionary Measures
In the case that table is updated this will force the previous rows to morph themselves to match the new columns, however if the new column added that does not have a default value then their may be some issues. And if a column is removed from the table then the old rows will have to have the column.